requirements.txt
flask
flask-login
werkzeug
psycopg2-binary
python-dotenv
gunicorn
apscheduler
pytz